requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3

# Email
fastapi-mail==1.4.1
jinja2==3.1.3

# AWS S3 for file storage
boto3==1.34.34
botocore==1.34.34

# Background tasks
celery==5.3.6
redis==5.0.1

# Validation and parsing
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator==2.1.0.post1

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Data processing
pandas==2.1.4
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Logging and monitoring
python-json-logger==2.0.7

# Date and time utilities
python-dateutil==2.8.2
pytz==2024.1

# Utilities
slugify==0.0.1
python-slugify==8.0.1
